+++ cmd/groups.go
@@ -6,7 +6,6 @@ import (
 
 	"fmt"
 	"os"
-	"sort"
 	"strings"
 
 	log "github.com/sirupsen/logrus"
@@ -26,7 +25,7 @@ func init() {
 	configCmd.AddCommand(groupCommand)
 	persistentFlags := groupCommand.PersistentFlags()
 	persistentFlags.BoolVarP(&listTargets, "list-targets", "t", false, "list targets in group")
-	persistentFlags.StringVarP(&group, "group", "g", "", "show only the specified group")
+	persistentFlags.StringVarP(&targetGroup, "group", "g", "", "show only the specified group")
 }
 
 func groups(_ *cobra.Command, _ []string) {
@@ -35,50 +34,42 @@ func groups(_ *cobra.Command, _ []string) {
 		log.Fatalf("Failed to load ospreyconfig file %s: %v", ospreyconfigFile, err)
 	}
 
-	groups := ospreyconfig.Groups()
-	sort.Strings(groups)
-	if len(groups) == 0 {
+	targets := client.GetTargets(ospreyconfig)
+	if !targets.HaveGroups() {
 		log.Error("There are no groups defined")
 		os.Exit(1)
 	}
 
-	var targetsByGroup map[string]map[string]*client.Osprey
-	if listTargets {
-		targetsByGroup = ospreyconfig.TargetsByGroup()
-	}
-
-	if group != "" {
-		if _, ok := targetsByGroup[group]; !ok {
-			log.Errorf("Group not found: %q", group)
+	var groups []client.Group
+	if targetGroup == "" {
+		groups = targets.Groups()
+	} else {
+		group, ok := targets.GetGroup(targetGroup)
+		if !ok {
+			log.Errorf("Group not found: %q", targetGroup)
 			os.Exit(1)
 		}
-		groups = []string{group}
+		log.Infof("Active group: %s", group.Name())
+		groups = []client.Group{group}
 	}
 
 	var outputLines []string
 	outputLines = append(outputLines, "Osprey groups:")
-	for _, groupName := range groups {
+	for _, group := range groups {
 		highlight := " "
-		if groupName == ospreyconfig.DefaultGroup {
+		if group.IsDefault() {
 			highlight = "*"
 		}
-		outputLines = append(outputLines, fmt.Sprintf("%s %s", highlight, groupName))
+		outputLines = append(outputLines, fmt.Sprintf("%s %s", highlight, group.Name()))
 
-		targets := targetsByGroup[groupName]
-		var targetNames []string
-		for targetName := range targets {
-			targetNames = append(targetNames, targetName)
-		}
-
-		sort.Strings(targetNames)
-		for _, name := range targetNames {
-			var aliases string
-			osprey := targets[name]
-			if len(osprey.Aliases) > 0 {
-				sort.Strings(osprey.Aliases)
-				aliases = fmt.Sprintf(" | %s", strings.Join(osprey.Aliases, ", "))
+		if listTargets {
+			for _, target := range group.Members() {
+				aliases := ""
+				if target.HasAliases() {
+					aliases = fmt.Sprintf(" | %s", strings.Join(target.Aliases(), ", "))
+				}
+				outputLines = append(outputLines, fmt.Sprintf("    %s%s", target.Name(), aliases))
 			}
-			outputLines = append(outputLines, fmt.Sprintf("    %s%s", name, aliases))
 		}
 	}
 	fmt.Println(strings.Join(outputLines, "\n"))