+++ CHANGELOG.md
@@ -1,6 +1,9 @@
+# Release 1.2.0
 - Handles `groups` for targets, allowing client commands to target a
   specific group.
-- Adds client commands: `config`, `config groups`.
+- Adds client commands: `config`, `config groups`, `config targets`.
+- All actions and outputs are now done against targets in alphabetical
+  order.
 
 # Release 1.1.0
 - Adds support for `certificate-authority-data` in client config.